Added parameters

+++ src/day_1.rs
@@ -1,7 +1,7 @@
-pub fn a(input: &str) -> i32 {
+pub fn a(input: &str) -> i64 {
   input
     .lines()
-    .map(|s| s.parse::<i32>().unwrap())
+    .map(|s| s.parse::<i64>().unwrap())
     .fold((0, None), |(count, prev_value), value| {
       (
         if prev_value.is_some() && value > prev_value.unwrap() {
@@ -14,15 +14,15 @@ pub fn a(input: &str) -> i32 {
     }).0
 }
 
-pub fn b(input: &str) -> i32 {
-  let rows: Vec<i32> = input
+pub fn b(input: &str) -> i64 {
+  let rows: Vec<i64> = input
     .lines()
     .map(|s| s.parse().unwrap())
     .collect();
 
   rows.windows(3)
     .map(|slice| slice.iter().sum())
-    .fold((0, None), |(count, prev_value), value: i32| {
+    .fold((0, None), |(count, prev_value), value: i64| {
       (
         if prev_value.is_some() && value > prev_value.unwrap() {
           count + 1
